refactor: extract gateway actions from AgentWorkspace
